refactor: EtfDetail에 inav/inavChangeAmount 타입 Long으로 변경

+++ frontend/WYE/app/src/main/java/com/d102/wye/domain/model/EtfDetail.kt
@@ -17,8 +17,8 @@ data class EtfDetail(
     val roe: Double,
     val aum: Long,
     val listingDate: String,
-    val inav: Double,
-    val inavChangeAmount: Double,
+    val inav: Long,
+    val inavChangeAmount: Long,
     val inavChangeRate: Double,
 )
